refactor: update imports to use views alias and add barrel exports
  - Update all imports from relative paths to use the 'views' alias for better consistency
  - Add barrel export files (index.ts) to Checkout module subdirectories for cleaner imports
  - Fix Vitest configuration to include views alias for test resolution
  - Standardize service imports to use named exports from barrel files
