+++ src/app/analytics/googleSheet.service.ts
@@ -1,7 +1,7 @@
 import { PriceWithTax } from '@internxt/sdk/dist/payments/types';
 import envService from 'app/core/services/env.service';
-import { CouponCodeData } from '../../views/Checkout/types';
-import { getProductAmount } from '../../views/Checkout/utils/getProductAmount';
+import { CouponCodeData } from 'views/Checkout/types';
+import { getProductAmount } from 'views/Checkout/utils';
 const WINTER_OFFSET_HOUR = 1;
 const TWO_EXTRA_HOURS_IN_MS = 2 * 60 * 60 * 1000;
