+++ src/app/analytics/impact.service.ts
@@ -9,7 +9,7 @@ import envService from 'app/core/services/env.service';
 import { PriceWithTax } from '@internxt/sdk/dist/payments/types';
 import { CouponCodeData } from 'views/Checkout/types';
 import { bytesToString } from 'app/drive/services/size.service';
-import { getProductAmount } from '../../views/Checkout/utils/getProductAmount';
+import { getProductAmount } from 'views/Checkout/utils';
 import { sendAddShoppersConversion } from './addShoppers.services';
 
 /**
