+++ src/app/store/slices/user/index.ts
@@ -22,7 +22,7 @@ import { workspacesActions } from '../../../store/slices/workspaces/workspacesSt
 import errorService from '../../../core/services/error.service';
 import { isTokenExpired } from '../../utils';
 import { refreshAvatar } from '../../../utils/avatar/avatarUtils';
-import { ProductService, UserTierFeatures } from '../../../../views/Checkout/services/products.service';
+import { ProductService, UserTierFeatures } from 'views/Checkout/services';
 import { t } from 'i18next';
 
 export interface UserState {
