+++ src/views/Checkout/components/CheckoutUserAuth.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable quotes */
-import { AuthMethodTypes } from '../../types';
+import { AuthMethodTypes } from '../types';
 import { FieldErrors, UseFormRegister } from 'react-hook-form';
 import { IFormValues } from 'app/core/types';
 import { Button, Avatar } from '@internxt/ui';
