+++ src/views/Checkout/components/index.ts
@@ -1,0 +1,8 @@
+export { AvailableCryptoCurrenciesDropdown } from './AvailableCryptoCurrenciesDropdown';
+export { CheckoutProductCard } from './CheckoutProductCard';
+export { CheckoutUserAuth } from './CheckoutUserAuth';
+export { CryptoPaymentDialog, CRYPTO_PAYMENT_DIALOG_KEY } from './CryptoPaymentDialog';
+export { HeaderComponent } from './Header';
+export { InputsComponent } from './InputsComponent';
+export { OptionalB2BDropdown } from './OptionalB2BDropdown';
+export { SelectSeatsComponent } from './SelectSeatsComponent';
