+++ src/views/Checkout/index.ts
@@ -1,0 +1,1 @@
+export * from './views';
