+++ src/views/Checkout/services/index.ts
@@ -1,0 +1,5 @@
+export { default as authCheckoutService } from './auth-checkout.service';
+export { default as checkoutService } from './checkout.service';
+export { default as currencyService } from './currency.service';
+export { default as paymentService } from './payment.service';
+export { fetchProducts, ProductService } from './products.service';
