+++ src/views/Checkout/store/index.ts
@@ -1,0 +1,2 @@
+export { checkoutReducer, initialStateForCheckout } from './checkoutReducer';
+export * from './types';
