+++ src/views/Checkout/utils/index.ts
@@ -1,0 +1,2 @@
+export { getProductAmount } from './getProductAmount';
+export { formatPrice } from './formatPrice';
