+++ src/views/Checkout/views/index.ts
@@ -1,0 +1,5 @@
+export { default as CheckoutCancelView } from './CheckoutCancelView/CheckoutCancelView';
+export { CheckoutSessionId } from './CheckoutSession/CheckoutSessionId';
+export { default as CheckoutSuccessView } from './CheckoutSuccessView/CheckoutSuccessView';
+export { default as PcCloudSuccess } from './CheckoutSuccessView/PcCloudSuccess';
+export { default as CheckoutViewWrapper } from './IntegratedCheckoutView/CheckoutViewWrapper';
