+++ src/views/NewSettings/components/Sections/Account/Plans/components/PlanSelection/PlanSelectionComponent.tsx
@@ -3,7 +3,7 @@ import PlanSelectionCard from './PlanSelectionCard';
 import { bytesToString } from 'app/drive/services/size.service';
 import { displayAmount } from '../../../../../../utils/planUtils';
 import { PlanSelectionCardSkeleton } from './skeletons/PlanSelectionCardSkeleton';
-import currencyService from '../../../../../../../views/Checkout/services/currency.service';
+import { currencyService } from 'views/Checkout/services';
 
 export interface UserSubscriptionProps {
   individual: UserSubscription | null;
