+++ src/views/NewSettings/components/Sections/Workspace/Billing/CancelSubscriptionModal.tsx
@@ -5,7 +5,7 @@ import sizeService from 'app/drive/services/size.service';
 import { FreeStoragePlan } from 'app/drive/types';
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import notificationsService, { ToastType } from 'app/notifications/services/notifications.service';
-import paymentService from 'views/Checkout/services/payment.service';
+import { paymentService } from 'views/Checkout/services';
 import { Button, Modal } from '@internxt/ui';
 import { useAppDispatch } from 'app/store/hooks';
 import { planThunks } from 'app/store/slices/plan';
