+++ src/views/NewSettings/services/plansApi.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest';
-import paymentService from 'views/Checkout/services/payment.service';
+import { paymentService } from 'views/Checkout/services';
 import envService from 'app/core/services/env.service';
 import { UserType } from '@internxt/sdk/dist/drive/payments/types/types';
 import { userLocation } from 'app/utils/userLocation';
@@ -13,8 +13,8 @@ vi.mock('@stripe/stripe-js', () => ({
   loadStripe: vi.fn(),
 }));
 
-vi.mock('views/Checkout/services/payment.service', () => ({
-  default: {
+vi.mock('views/Checkout/services', () => ({
+  paymentService: {
     getPrices: vi.fn(),
   },
 }));
